--- a/.github/workflows/lint-provisioning-scripts.yml
+++ b/.github/workflows/lint-provisioning-scripts.yml
@@ -16,6 +16,9 @@ jobs:
     runs-on: ubuntu-20.04
 
     steps:
+
+      - uses: actions/checkout@v2
+
       - name: install shellcheck
         run: sudo apt-get install shellcheck