--- a/docker-compose.yml
+++ b/docker-compose.yml
@@ -19,7 +19,7 @@ services:
   chrome:
     container_name: "edx.${COMPOSE_PROJECT_NAME:-devstack}.chrome"
     hostname: chrome.devstack.edx
-    image: edxops/chrome:${OPENEDX_RELEASE:-latest}
+    image: ${_CHROME_IMAGE}
     shm_size: 2g
     networks:
       default:
@@ -54,7 +54,7 @@ services:
   devpi:
     container_name: "edx.${COMPOSE_PROJECT_NAME:-devstack}.devpi"
     hostname: devpi.devstack.edx
-    image: edxops/devpi:${OPENEDX_RELEASE:-latest}
+    image: ${_DEVPI_IMAGE}
     networks:
       default:
         aliases:
@@ -121,7 +121,7 @@ services:
   firefox:
     container_name: "edx.${COMPOSE_PROJECT_NAME:-devstack}.firefox"
     hostname: firefox.devstack.edx
-    image: edxops/firefox:${OPENEDX_RELEASE:-latest}
+    image: ${_FIREFOX_IMAGE}
     shm_size: 2g
     networks:
       default:
@@ -300,7 +300,7 @@ services:
       SOCIAL_AUTH_EDX_OIDC_URL_ROOT: http://edx.devstack.lms:18000/oauth2
       ENABLE_DJANGO_TOOLBAR: 1
       DJANGO_WATCHMAN_TIMEOUT: 30
-    image: edxops/credentials:${OPENEDX_RELEASE:-latest}
+    image: ${_CREDENTIALS_IMAGE}
     networks:
       default:
         aliases:
@@ -326,7 +326,7 @@ services:
       TEST_ELASTICSEARCH_URL: "edx.devstack.elasticsearch710"
       ENABLE_DJANGO_TOOLBAR: 1
       DJANGO_WATCHMAN_TIMEOUT: 30
-    image: edxops/discovery:${OPENEDX_RELEASE:-latest}
+    image: ${_DISCOVERY_IMAGE}
     networks:
       default:
         aliases:
@@ -351,7 +351,7 @@ services:
     environment:
       DJANGO_WATCHMAN_TIMEOUT: 30
       ENABLE_DJANGO_TOOLBAR: 1
-    image: edxops/ecommerce:${OPENEDX_RELEASE:-latest}
+    image: ${_ECOMMERCE_IMAGE}
     networks:
       default:
         aliases:
@@ -369,7 +369,7 @@ services:
       - elasticsearch710
       - lms
       - mysql57
-    image: edxops/notes:${OPENEDX_RELEASE:-latest}
+    image: ${_NOTES_IMAGE}
     networks:
       default:
         aliases:
@@ -396,7 +396,7 @@ services:
       - memcached
       - mongo
       - elasticsearch710
-    image: edxops/forum:${OPENEDX_RELEASE:-latest}
+    image: ${_FORUM_IMAGE}
     stdin_open: true
     tty: true
     networks:
@@ -428,7 +428,7 @@ services:
       EDXAPP_TEST_MONGO_HOST: edx.devstack.mongo
       NO_PYTHON_UNINSTALL: 1
       DJANGO_WATCHMAN_TIMEOUT: 30
-    image: edxops/edxapp:${OPENEDX_RELEASE:-latest}
+    image: ${_LMS_IMAGE}
     networks:
       default:
         aliases:
@@ -462,7 +462,7 @@ services:
       DB_PASSWORD: password
       LMS_HOST: http://localhost:18000
       DJANGO_SETTINGS_MODULE: analytics_dashboard.settings.devstack
-    image: edxops/insights:${OPENEDX_RELEASE:-latest}
+    image: ${_INSIGHTS_IMAGE}
     networks:
       default:
         aliases:
@@ -473,7 +473,7 @@ services:
       - /edx/var/insights/
 
   analyticsapi:
-    image: edxops/analytics_api:${OPENEDX_RELEASE:-latest}
+    image: ${_ANALYTICSAPI_IMAGE}
     container_name: edx.devstack.analyticsapi
     hostname: analyticsapi
     depends_on:
@@ -522,7 +522,7 @@ services:
       CELERY_BROKER_VHOST: 10
       CELERY_BROKER_PASSWORD: password
       DJANGO_WATCHMAN_TIMEOUT: 30
-    image: edxops/registrar:${OPENEDX_RELEASE:-latest}
+    image: ${_REGISTRAR_IMAGE}
     networks:
       default:
         aliases:
@@ -556,7 +556,7 @@ services:
       CELERY_BROKER_VHOST: 10
       CELERY_BROKER_PASSWORD: password
       DJANGO_WATCHMAN_TIMEOUT: 30
-    image: edxops/registrar:${OPENEDX_RELEASE:-latest}
+    image: ${_REGISTRAR_IMAGE}
     networks:
       default:
         aliases:
@@ -587,7 +587,7 @@ services:
       EDXAPP_TEST_MONGO_HOST: edx.devstack.mongo
       NO_PYTHON_UNINSTALL: 1
       DJANGO_WATCHMAN_TIMEOUT: 30
-    image: edxops/edxapp:${OPENEDX_RELEASE:-latest}
+    image: ${_CMS_IMAGE}
     networks:
       default:
         aliases:
@@ -603,7 +603,7 @@ services:
 
   xqueue:
     container_name: "edx.${COMPOSE_PROJECT_NAME:-devstack}.xqueue"
-    image: edxops/xqueue:${OPENEDX_RELEASE:-latest}
+    image: ${_XQUEUE_IMAGE}
     command: bash -c 'source /edx/app/xqueue/xqueue_env && while true; do python /edx/app/xqueue/xqueue/manage.py runserver 0.0.0.0:18040 ; sleep 2; done'
     volumes:
       - ${DEVSTACK_WORKSPACE}/xqueue:/edx/app/xqueue/xqueue
@@ -618,7 +618,7 @@ services:
 
   xqueue_consumer:
     container_name: "edx.${COMPOSE_PROJECT_NAME:-devstack}.xqueue_consumer"
-    image: edxops/xqueue:${OPENEDX_RELEASE:-latest}
+    image: ${_XQUEUE_IMAGE}
     command: bash -c 'source /edx/app/xqueue/xqueue_env && while true; do python /edx/app/xqueue/xqueue/manage.py run_consumer ; sleep 2; done'
     volumes:
       - ${DEVSTACK_WORKSPACE}/xqueue:/edx/app/xqueue/xqueue