--- a/scripts/colors.sh
+++ b/scripts/colors.sh
@@ -1,4 +1,8 @@
+#!/usr/bin/env bash
 # Source this file to get color variables
+
+# Disable unused variable warning.
+# shellcheck disable=SC2034
 
 RED='\033[0;31m'
 GREEN='\033[0;32m'