@syedimranhassan 2022-01-06
@kdmccormick The GH action added in this PR is showing up in other PR too and not triggering because it's not merged yet.
Here is my PR [871](https://github.com/openedx/devstack/pull/871)  and this GH action not allowing me to merge my PR.