@kdmccormick 2022-01-21
In light of the kickoff of [work to replace devstack with tutor](https://github.com/openedx/platform-roadmap/issues/11), I don't feel that it's worth our while any more to review this and roll it out. We can always reopen this later if it became useful again.